pyproject.toml
[project]
name = "jarvis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.43",
    "google-genai>=1.41.0",
    "langchain>=0.3.27",
    "loguru>=0.7.3",
    "minio>=7.2.18",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "pymupdf>=1.26.4",
    "pymupdf4llm>=0.0.27",
    "python-dotenv>=1.1.1",
    "qdrant-client[fastembed]>=1.15.1",
    "rich>=14.1.0",
    "sentence-transformers>=5.1.1",
    "sqlalchemy-utils>=0.42.0",
    "typer>=0.19.2",
    "zenml[local,server]==0.90.0",
]

[project.scripts]
jarvis = "jarvis.main:app"
